+++ backend/etest/src/main/java/edu/fzu/etest/service/QuestionService.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Service
 public interface QuestionService {
     void add(Question question);//添加问题
-    List<Question> getQuestionsByType(int aid, int type); //获取不同类型的问题
+    List<Question> getQuestionsByType(long aid, int type); //获取不同类型的问题
 }
